examples/text_to_agent/basic_chat/gemini_live/pyproject.toml
[project]
name = "basic-chat-gemini-live"
version = "0.1.0"
description = "Basic chat agent using Gemini Live"
dependencies = [
    "cartesia-line",
    "google-genai",
    "loguru",
    "fastapi",
    "uvicorn",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
